requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
yfinance>=0.2.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
gymnasium>=0.29.0
requests>=2.31.0
lxml>=4.9.0
tensorflow